% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{The function \code{fars_read} is used to read the data from a .csv file.
It assumes that the input file is a is in a csv formar. In cases the input 
file name is incorrect, it outputs an error.}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{a name of the input file in the current folder or a full path to the file}
}
\value{
The function returns a data frame in a tibble format as a result of an application of tbl_df function
}
\description{
The function \code{fars_read} is used to read the data from a .csv file.
It assumes that the input file is a is in a csv formar. In cases the input 
file name is incorrect, it outputs an error.
}
\note{
An error may result if the file is in a format that cannot be correctly read by the read_csv function
}
\examples{
fars_read("data/accident_2013.csv.bz2")

}
