% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FARfunctions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{The function \code{fars_read_years} reads the values of month and year from individual reports
based on a vector of year values and generates a list of data frames for further processing by another function.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{a vector of integer year values}
}
\value{
The function returns a list of data frames, each of which consists of Month and year columns
}
\description{
The function \code{fars_read_years} reads the values of month and year from individual reports
based on a vector of year values and generates a list of data frames for further processing by another function.
}
\note{
The function deals effectively with inappropriate input values.
}
\examples{
\dontrun{fars_read_years(c(2014, 2015))}

}
