% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FARfunctions.R
\name{make_filename}
\alias{make_filename}
\title{The function \code{make_filename} automatically generates a name for a new file based on the input year.}
\usage{
make_filename(year)
}
\arguments{
\item{year}{An integer value for the year in a user defined format, e.g. 17 or 2017}
}
\value{
The function returns a string for the file name in the format "accident_%d.csv.bz2", where %d
is a placeholder for the integer value of the year
}
\description{
The function \code{make_filename} automatically generates a name for a new file based on the input year.
}
